requirements.txt
colorama==0.4.6
cycler==0.11.0
docopt==0.6.2
fastremap==1.14.1
fill-voids==2.0.6
fonttools==4.38.0
imageio==2.31.2
kiwisolver==1.4.5
lungmask==0.2.16
matplotlib==3.5.3
more-itertools==9.1.0
natsort==8.4.0
networkx==2.6.3
numpy==1.21.6
packaging==23.2
pandas==1.3.5
Pillow==9.5.0
pydicom==2.4.4
pykwalify==1.8.0
pyparsing==3.1.1
pyradiomics==3.0.1
python-dateutil==2.8.2
pytz==2024.1
PyWavelets==1.3.0
PyYAML==6.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-image==0.19.3
scipy==1.7.3
SimpleITK==2.2.1
six==1.16.0
tifffile==2021.11.2
torch==1.13.1
tqdm==4.66.2
typing_extensions==4.7.1
wincertstore==0.2
